+++ magik-squid/src/main/java/nl/ramsolutions/sw/magik/analysis/typing/TypeKeeper.java
@@ -274,4 +274,9 @@ public class TypeKeeper implements ITypeKeeper {
         this.conditions.remove(name);
     }
 
+    @Override
+    public Collection<BinaryOperator> getBinaryOperators() {
+        return Collections.unmodifiableCollection(this.binaryOperators);
+    }
+
 }