+++ magik-squid/src/main/java/nl/ramsolutions/sw/magik/analysis/typing/io/InstPackage.java
@@ -1,0 +1,22 @@
+package nl.ramsolutions.sw.magik.analysis.typing.io;
+
+/**
+ * Package instructions.
+ */
+@SuppressWarnings("checkstyle:JavadocVariable")
+enum InstPackage {
+
+    NAME("name"),
+    USES("uses");
+
+    private final String instruction;
+
+    InstPackage(final String instruction) {
+        this.instruction = instruction;
+    }
+
+    public String getValue() {
+        return this.instruction;
+    }
+
+}