+++ magik-squid/src/main/java/nl/ramsolutions/sw/magik/analysis/typing/io/JsonTypeKeeperReader.java
@@ -1,4 +1,4 @@
-package nl.ramsolutions.sw.magik.analysis.typing;
+package nl.ramsolutions.sw.magik.analysis.typing.io;
 
 import java.io.BufferedReader;
 import java.io.File;
@@ -12,6 +12,10 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 import nl.ramsolutions.sw.magik.analysis.Location;
+import nl.ramsolutions.sw.magik.analysis.typing.BinaryOperator;
+import nl.ramsolutions.sw.magik.analysis.typing.ITypeKeeper;
+import nl.ramsolutions.sw.magik.analysis.typing.TypeKeeper;
+import nl.ramsolutions.sw.magik.analysis.typing.TypeReader;
 import nl.ramsolutions.sw.magik.analysis.typing.types.AbstractType;
 import nl.ramsolutions.sw.magik.analysis.typing.types.AliasType;
 import nl.ramsolutions.sw.magik.analysis.typing.types.Condition;
@@ -61,12 +65,7 @@ public final class JsonTypeKeeperReader {
              BufferedReader bufferedReader = new BufferedReader(fileReader)) {
             String line = bufferedReader.readLine();
             while (line != null) {
-                try {
-                    this.processLine(line);
-                } catch (final RuntimeException exception) {
-                    LOGGER.error("Error parsing line {}, line data: {}", lineNo, line);
-                    LOGGER.error(exception.getMessage(), exception);
-                }
+                this.processLineSafe(lineNo, line);
 
                 ++lineNo;
                 line = bufferedReader.readLine();
@@ -78,6 +77,16 @@ public final class JsonTypeKeeperReader {
         }
     }
 
+    @SuppressWarnings("checkstyle:IllegalCatch")
+    private void processLineSafe(int lineNo, String line) {
+        try {
+            this.processLine(line);
+        } catch (final RuntimeException exception) {
+            LOGGER.error("Error parsing line {}, line data: {}", lineNo, line);
+            LOGGER.error(exception.getMessage(), exception);
+        }
+    }
+
     private void processLine(final String line) {
         if (line.startsWith("//")) {
             // Ignore comments.
@@ -86,33 +95,34 @@ public final class JsonTypeKeeperReader {
 
         final JSONTokener tokener = new JSONTokener(line);
         final JSONObject obj = new JSONObject(tokener);
-        final String instruction = obj.getString("instruction");
+        final String instructionStr = obj.getString(InstInstruction.INSTRUCTION.getValue());
+        final InstInstruction instruction = InstInstruction.fromValue(instructionStr);
         switch (instruction) {
-            case "package":
+            case PACKAGE:
                 this.handlePackage(obj);
                 break;
 
-            case "type":
+            case TYPE:
                 this.handleType(obj);
                 break;
 
-            case "global":
+            case GLOBAL:
                 this.handleGlobal(obj);
                 break;
 
-            case "method":
+            case METHOD:
                 this.handleMethod(obj);
                 break;
 
-            case "procedure":
+            case PROCEDURE:
                 this.handleProcedure(obj);
                 break;
 
-            case "condition":
+            case CONDITION:
                 this.handleCondition(obj);
                 break;
 
-            case "binary_operator":
+            case BINARY_OPERATOR:
                 this.handleBinaryOperator(obj);
                 break;
 
@@ -122,10 +132,10 @@ public final class JsonTypeKeeperReader {
     }
 
     private void handlePackage(final JSONObject instruction) {
-        final String pakkageName = instruction.getString("name");
+        final String pakkageName = instruction.getString(InstPackage.NAME.getValue());
         final Package pakkage = this.ensurePackage(pakkageName);
 
-        instruction.getJSONArray("uses").forEach(useObj -> {
+        instruction.getJSONArray(InstPackage.USES.getValue()).forEach(useObj -> {
             final String use = (String) useObj;
             this.ensurePackage(use);
             pakkage.addUse(use);
@@ -133,46 +143,48 @@ public final class JsonTypeKeeperReader {
     }
 
     private void handleType(final JSONObject instruction) {
-        final String typeFormat = instruction.getString("type_format");
-        final String name = instruction.getString("type_name");
+        final String typeFormat = instruction.getString(InstType.TYPE_FORMAT.getValue());
+        final String name = instruction.getString(InstType.TYPE_NAME.getValue());
         final TypeString typeString = TypeStringParser.parseTypeString(name);
         final MagikType type;
         if (this.typeKeeper.getType(typeString) != UndefinedType.INSTANCE) {
             type = (MagikType) this.typeKeeper.getType(typeString);
-        } else if ("intrinsic".equals(typeFormat)) {
+        } else if (InstType.INTRINSIC.getValue().equals(typeFormat)) {
             type = new MagikType(this.typeKeeper, Sort.INTRINSIC, typeString);
-        } else if ("slotted".equals(typeFormat)) {
+        } else if (InstType.SLOTTED.getValue().equals(typeFormat)) {
             type = new MagikType(this.typeKeeper, Sort.SLOTTED, typeString);
-        } else if ("indexed".equals(typeFormat)) {
+        } else if (InstType.INDEXED.getValue().equals(typeFormat)) {
             type = new MagikType(this.typeKeeper, Sort.INDEXED, typeString);
+        } else if (InstType.OBJECT.getValue().equals(typeFormat)) {
+            type = new MagikType(this.typeKeeper, Sort.OBJECT, typeString);
         } else {
             throw new InvalidParameterException("Unknown type: " + typeFormat);
         }
 
-        if (instruction.has("doc")) {
-            final String doc = instruction.getString("doc");
+        if (instruction.has(InstType.DOC.getValue())) {
+            final String doc = instruction.getString(InstType.DOC.getValue());
             type.setDoc(doc);
         }
 
-        instruction.getJSONArray("slots").forEach(slotObj -> {
+        instruction.getJSONArray(InstType.SLOTS.getValue()).forEach(slotObj -> {
             final JSONObject slot = (JSONObject) slotObj;
-            final String slotName = slot.getString("name");
-            final String slotTypeName = slot.getString("type_name");
+            final String slotName = slot.getString(InstType.SLOT_NAME.getValue());
+            final String slotTypeName = slot.getString(InstType.SLOT_TYPE_NAME.getValue());
             final TypeString slotTypeString = TypeStringParser.parseTypeString(slotTypeName);
             type.addSlot(null, slotName, slotTypeString);
         });
 
-        if (instruction.has("generics")) {
-            instruction.getJSONArray("generics").forEach(genericObj -> {
+        if (instruction.has(InstType.GENERICS.getValue())) {
+            instruction.getJSONArray(InstType.GENERICS.getValue()).forEach(genericObj -> {
                 final JSONObject generic = (JSONObject) genericObj;
-                final String genericName = generic.getString("name");
-                final String genericDoc = generic.getString("doc");
+                final String genericName = generic.getString(InstType.GENERIC_NAME.getValue());
+                final String genericDoc = generic.getString(InstType.GENERIC_DOC.getValue());
                 final GenericDeclaration genericDeclaration = type.addGeneric(null, genericName);
                 genericDeclaration.setDoc(genericDoc);
             });
         }
 
-        final JSONArray parents = instruction.getJSONArray("parents");
+        final JSONArray parents = instruction.getJSONArray(InstType.PARENTS.getValue());
         parents.forEach(parentObj -> {
             final String parentStr = (String) parentObj;
             final TypeString parentRef = TypeString.ofIdentifier(parentStr, SW_PAKKAGE);
@@ -181,7 +193,7 @@ public final class JsonTypeKeeperReader {
     }
 
     private void handleGlobal(final JSONObject instruction) {
-        final String name = instruction.getString("name");
+        final String name = instruction.getString(InstGlobal.NAME.getValue());
         final TypeString typeString = TypeString.ofIdentifier(name, TypeString.DEFAULT_PACKAGE);
         final AbstractType type = this.typeKeeper.getType(typeString);
         if (type != UndefinedType.INSTANCE) {
@@ -190,7 +202,7 @@ public final class JsonTypeKeeperReader {
             return;
         }
 
-        final String typeName = instruction.getString("type_name");
+        final String typeName = instruction.getString(InstGlobal.TYPE_NAME.getValue());
         final TypeString aliasedRef = TypeStringParser.parseTypeString(typeName);
 
         final AliasType global = new AliasType(this.typeKeeper, typeString, aliasedRef);
@@ -198,7 +210,7 @@ public final class JsonTypeKeeperReader {
     }
 
     private void handleMethod(final JSONObject instruction) {
-        final String typeName = instruction.getString("type_name");
+        final String typeName = instruction.getString(InstMethod.TYPE_NAME.getValue());
         final TypeString typeRef = TypeString.ofIdentifier(typeName, TypeString.DEFAULT_PACKAGE);
         final AbstractType abstractType = this.typeKeeper.getType(typeRef);
         if (abstractType == UndefinedType.INSTANCE) {
@@ -207,9 +219,9 @@ public final class JsonTypeKeeperReader {
         final MagikType type = (MagikType) abstractType;
 
         // What if method already exists? Might be improved by taking the line into account as well.
-        final String methodName = instruction.getString("method_name");
-        final Location location = instruction.get("source_file") != JSONObject.NULL
-            ? new Location(Path.of(instruction.getString("source_file")).toUri())
+        final String methodName = instruction.getString(InstMethod.METHOD_NAME.getValue());
+        final Location location = instruction.get(InstMethod.SOURCE_FILE.getValue()) != JSONObject.NULL
+            ? new Location(Path.of(instruction.getString(InstMethod.SOURCE_FILE.getValue())).toUri())
             : null;
         final boolean isAlreadyKnown = type.getLocalMethods(methodName).stream()
             .anyMatch(method ->
@@ -220,13 +232,14 @@ public final class JsonTypeKeeperReader {
             return;
         }
 
-        final JSONArray modifiersArray = (JSONArray) instruction.getJSONArray("modifiers");
+        final JSONArray modifiersArray = instruction.getJSONArray(InstMethod.MODIFIERS.getValue());
         final EnumSet<Method.Modifier> modifiers = StreamSupport.stream(modifiersArray.spliterator(), false)
             .map(String.class::cast)
             .map(String::toUpperCase)
             .map(Method.Modifier::valueOf)
             .collect(Collectors.toCollection(() -> EnumSet.noneOf(Method.Modifier.class)));
-        final List<Parameter> parameters = this.parseParameters(instruction.getJSONArray("parameters"));
+        final List<Parameter> parameters =
+            this.parseParameters(instruction.getJSONArray(InstMethod.PARAMETERS.getValue()));
         final Parameter assignmentParameter;
         if (methodName.contains("<<")) {
             int lastIdx = parameters.size() - 1;
@@ -235,22 +248,24 @@ public final class JsonTypeKeeperReader {
         } else {
             assignmentParameter = null;
         }
-        final ExpressionResultString result =  this.parseExpressionResultString(instruction.get("return_types"));
-        final ExpressionResultString loopResult = this.parseExpressionResultString(instruction.get("loop_types"));
-        final String methodDoc = instruction.get("doc") != JSONObject.NULL
-            ? instruction.getString("doc")
+        final ExpressionResultString result =
+            this.parseExpressionResultString(instruction.get(InstMethod.RETURN_TYPES.getValue()));
+        final ExpressionResultString loopResult =
+            this.parseExpressionResultString(instruction.get(InstMethod.LOOP_TYPES.getValue()));
+        final String methodDoc = instruction.get(InstMethod.DOC.getValue()) != JSONObject.NULL
+            ? instruction.getString(InstMethod.DOC.getValue())
             : null;
         type.addMethod(location, modifiers, methodName, parameters, assignmentParameter, methodDoc, result, loopResult);
     }
 
     private void handleCondition(final JSONObject instruction) {
-        final String name = instruction.getString("name");
-        final String doc = instruction.getString("doc");
-        final String parent = instruction.getString("parent");
-        final Location location = instruction.get("source_file") != JSONObject.NULL
-            ? new Location(Path.of(instruction.getString("source_file")).toUri())
+        final String name = instruction.getString(InstCondition.NAME.getValue());
+        final String doc = instruction.getString(InstCondition.DOC.getValue());
+        final String parent = instruction.getString(InstCondition.PARENT.getValue());
+        final Location location = instruction.get(InstCondition.SOURCE_FILE.getValue()) != JSONObject.NULL
+            ? new Location(Path.of(instruction.getString(InstCondition.SOURCE_FILE.getValue())).toUri())
             : null;
-        final JSONArray dataNameListArray = (JSONArray) instruction.getJSONArray("data_name_list");
+        final JSONArray dataNameListArray = instruction.getJSONArray(InstCondition.DATA_NAME_LIST.getValue());
         final List<String> dataNameList = StreamSupport.stream(dataNameListArray.spliterator(), false)
             .map(String.class::cast)
             .collect(Collectors.toList());
@@ -260,12 +275,12 @@ public final class JsonTypeKeeperReader {
 
     private void handleBinaryOperator(final JSONObject instruction) {
         final BinaryOperator.Operator operator =
-            BinaryOperator.Operator.valueFor(instruction.getString("operator"));
-        final String lhsTypeStr = instruction.getString("lhs_type");
+            BinaryOperator.Operator.valueFor(instruction.getString(InstBinaryOperator.OPERATOR.getValue()));
+        final String lhsTypeStr = instruction.getString(InstBinaryOperator.LHS_TYPE.getValue());
         final TypeString lhsTypeRef = TypeStringParser.parseTypeString(lhsTypeStr);
-        final String rhsTypeStr = instruction.getString("rhs_type");
+        final String rhsTypeStr = instruction.getString(InstBinaryOperator.RHS_TYPE.getValue());
         final TypeString rhsTypeRef = TypeStringParser.parseTypeString(rhsTypeStr);
-        final String returnTypeStr = instruction.getString("return_type");
+        final String returnTypeStr = instruction.getString(InstBinaryOperator.RETURN_TYPE.getValue());
         final TypeString resultTypeRef = TypeStringParser.parseTypeString(returnTypeStr);
         final BinaryOperator binaryOperator = new BinaryOperator(operator, lhsTypeRef, rhsTypeRef, resultTypeRef);
         this.typeKeeper.addBinaryOperator(binaryOperator);
@@ -275,11 +290,13 @@ public final class JsonTypeKeeperReader {
         return StreamSupport.stream(parametersArray.spliterator(), false)
             .map(parameterThing -> {
                 final JSONObject parameterObj = (JSONObject) parameterThing;
-                final String name = parameterObj.getString("name");
-                final Parameter.Modifier modifier = parameterObj.get("modifier") != JSONObject.NULL
-                    ? Parameter.Modifier.valueOf(parameterObj.getString("modifier").toUpperCase())
+                final String name = parameterObj.getString(InstParameter.NAME.getValue());
+                final Parameter.Modifier modifier =
+                    parameterObj.get(InstParameter.MODIFIER.getValue()) != JSONObject.NULL
+                    ? Parameter.Modifier.valueOf(
+                        parameterObj.getString(InstParameter.MODIFIER.getValue()).toUpperCase())
                     : Parameter.Modifier.NONE;
-                final String typeName = parameterObj.getString("type_name");
+                final String typeName = parameterObj.getString(InstParameter.TYPE_NAME.getValue());
                 final TypeString typeString = TypeStringParser.parseTypeString(typeName);
                 final AbstractType type = this.typeParser.parseTypeString(typeString);
                 return new Parameter(name, modifier, type);
@@ -309,23 +326,26 @@ public final class JsonTypeKeeperReader {
     }
 
     private void handleProcedure(final JSONObject instruction) {
-        final JSONArray modifiersArray = (JSONArray) instruction.getJSONArray("modifiers");
+        final JSONArray modifiersArray = instruction.getJSONArray(InstProcedure.MODIFIERS.getValue());
         final EnumSet<Method.Modifier> modifiers = StreamSupport.stream(modifiersArray.spliterator(), false)
             .map(String.class::cast)
             .map(String::toUpperCase)
             .map(Method.Modifier::valueOf)
             .collect(Collectors.toCollection(() -> EnumSet.noneOf(Method.Modifier.class)));
-        final Location location = instruction.get("source_file") != JSONObject.NULL
-            ? new Location(Path.of(instruction.getString("source_file")).toUri())
+        final Location location = instruction.get(InstProcedure.SOURCE_FILE.getValue()) != JSONObject.NULL
+            ? new Location(Path.of(instruction.getString(InstProcedure.SOURCE_FILE.getValue())).toUri())
             : null;
-        final List<Parameter> parameters = this.parseParameters(instruction.getJSONArray("parameters"));
-        final ExpressionResultString resultStr = this.parseExpressionResultString(instruction.get("return_types"));
-        final ExpressionResultString loopResultStr = this.parseExpressionResultString(instruction.get("loop_types"));
+        final List<Parameter> parameters =
+            this.parseParameters(instruction.getJSONArray(InstProcedure.PARAMETERS.getValue()));
+        final ExpressionResultString resultStr =
+            this.parseExpressionResultString(instruction.get(InstProcedure.RETURN_TYPES.getValue()));
+        final ExpressionResultString loopResultStr =
+            this.parseExpressionResultString(instruction.get(InstProcedure.LOOP_TYPES.getValue()));
 
-        final String procedureName = instruction.getString("procedure_name");
+        final String procedureName = instruction.getString(InstProcedure.PROCEDURE_NAME.getValue());
         final MagikType procedureType = (MagikType) this.typeKeeper.getType(SW_PROCEDURE_REF);
-        final String methodDoc = instruction.get("doc") != JSONObject.NULL
-            ? instruction.getString("doc")
+        final String methodDoc = instruction.get(InstProcedure.DOC.getValue()) != JSONObject.NULL
+            ? instruction.getString(InstProcedure.DOC.getValue())
             : null;
         final ProcedureInstance instance = new ProcedureInstance(
             procedureType,
@@ -338,7 +358,7 @@ public final class JsonTypeKeeperReader {
             loopResultStr);
 
         // Create alias to instance.
-        final String name = instruction.getString("name");
+        final String name = instruction.getString(InstProcedure.NAME.getValue());
         final TypeString typeString = TypeString.ofIdentifier(name, TypeString.DEFAULT_PACKAGE);
         final AbstractType type = this.typeKeeper.getType(typeString);
         if (type != UndefinedType.INSTANCE) {